Remove Telemetry module; transfer CrashHandler to Diagnostics module
