+++ src/appshell/view/preferences/generalpreferencesmodel.h
@@ -32,7 +32,6 @@
 #include "languages/ilanguagesservice.h"
 #include "shortcuts/ishortcutsconfiguration.h"
 #include "project/iprojectconfiguration.h"
-#include "telemetry/itelemetryconfiguration.h"
 
 namespace mu::appshell {
 class GeneralPreferencesModel : public QObject, public async::Asyncable
@@ -44,7 +43,6 @@ class GeneralPreferencesModel : public QObject, public async::Asyncable
     INJECT(appshell, languages::ILanguagesService, languagesService)
     INJECT(appshell, shortcuts::IShortcutsConfiguration, shortcutsConfiguration)
     INJECT(appshell, project::IProjectConfiguration, projectConfiguration)
-    INJECT(appshell, telemetry::ITelemetryConfiguration, telemetryConfiguration)
 
     Q_PROPERTY(QVariantList languages READ languages NOTIFY languagesChanged)
     Q_PROPERTY(QString currentLanguageCode READ currentLanguageCode WRITE setCurrentLanguageCode NOTIFY currentLanguageCodeChanged)
@@ -52,7 +50,6 @@ class GeneralPreferencesModel : public QObject, public async::Asyncable
     Q_PROPERTY(QStringList keyboardLayouts READ keyboardLayouts CONSTANT)
     Q_PROPERTY(QString currentKeyboardLayout READ currentKeyboardLayout WRITE setCurrentKeyboardLayout NOTIFY currentKeyboardLayoutChanged)
 
-    Q_PROPERTY(bool isTelemetryAllowed READ isTelemetryAllowed WRITE setIsTelemetryAllowed NOTIFY isTelemetryAllowedChanged)
     Q_PROPERTY(bool isAutoSaveEnabled READ isAutoSaveEnabled WRITE setAutoSaveEnabled NOTIFY autoSaveEnabledChanged)
     Q_PROPERTY(int autoSaveInterval READ autoSaveInterval WRITE setAutoSaveInterval NOTIFY autoSaveIntervalChanged)
     Q_PROPERTY(bool isOSCRemoteControl READ isOSCRemoteControl WRITE setIsOSCRemoteControl NOTIFY isOSCRemoteControlChanged)
@@ -70,7 +67,6 @@ public:
     QStringList keyboardLayouts() const;
     QString currentKeyboardLayout() const;
 
-    bool isTelemetryAllowed() const;
     bool isAutoSaveEnabled() const;
     int autoSaveInterval() const;
     bool isOSCRemoteControl() const;
@@ -79,7 +75,6 @@ public:
 public slots:
     void setCurrentLanguageCode(const QString& currentLanguageCode);
     void setCurrentKeyboardLayout(const QString& keyboardLayout);
-    void setIsTelemetryAllowed(bool isTelemetryAllowed);
     void setAutoSaveEnabled(bool enabled);
     void setAutoSaveInterval(int minutes);
     void setIsOSCRemoteControl(bool isOSCRemoteControl);
@@ -89,7 +84,6 @@ signals:
     void languagesChanged(QVariantList languages);
     void currentLanguageCodeChanged(QString currentLanguageCode);
     void currentKeyboardLayoutChanged();
-    void isTelemetryAllowedChanged(bool isTelemetryAllowed);
     void autoSaveEnabledChanged(bool enabled);
     void autoSaveIntervalChanged(int minutes);
     void isOSCRemoteControlChanged(bool isOSCRemoteControl);
