+++ src/diagnostics/CMakeLists.txt
@@ -31,6 +31,8 @@ set(MODULE_SRC
     ${CMAKE_CURRENT_LIST_DIR}/idiagnosticspathsregister.h
     ${CMAKE_CURRENT_LIST_DIR}/iengravingelementsprovider.h
 
+    ${CMAKE_CURRENT_LIST_DIR}/internal/diagnosticsconfiguration.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/internal/diagnosticsconfiguration.h
     ${CMAKE_CURRENT_LIST_DIR}/internal/diagnosticsactions.cpp
     ${CMAKE_CURRENT_LIST_DIR}/internal/diagnosticsactions.h
     ${CMAKE_CURRENT_LIST_DIR}/internal/diagnosticsactionscontroller.cpp
@@ -61,7 +63,51 @@ set(MODULE_SRC
 
     ${CMAKE_CURRENT_LIST_DIR}/view/engraving/engravingelementsmodel.cpp
     ${CMAKE_CURRENT_LIST_DIR}/view/engraving/engravingelementsmodel.h
+
+    ${CMAKE_CURRENT_LIST_DIR}/devtools/crashhandlerdevtoolsmodel.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/devtools/crashhandlerdevtoolsmodel.h
     )
+
+# --- Crashpad ---
+# Not building with MinGW, so turned off for MinGW
+include(GetCompilerInfo)
+set(BUILD_CRASHPAD_CLIENT ON)
+if (CC_IS_MINGW)
+    set(BUILD_CRASHPAD_CLIENT OFF)
+endif(CC_IS_MINGW)
+
+if (BUILD_CRASHPAD_CLIENT)
+    set(MODULE_SRC ${MODULE_SRC}
+        ${CMAKE_CURRENT_LIST_DIR}/internal/crashhandler/crashhandler.cpp
+        ${CMAKE_CURRENT_LIST_DIR}/internal/crashhandler/crashhandler.h
+    )
+
+    set(MODULE_INCLUDE ${MODULE_INCLUDE}
+        ${PROJECT_SOURCE_DIR}/thirdparty/google_crashpad_client
+        ${PROJECT_SOURCE_DIR}/thirdparty/google_crashpad_client/third_party/mini_chromium/mini_chromium
+    )
+
+    set(MODULE_DEF ${MODULE_DEF}
+        -DBUILD_CRASHPAD_CLIENT
+    )
+
+    add_subdirectory(${PROJECT_SOURCE_DIR}/thirdparty/google_crashpad_client google_crashpad_client)
+
+    set(MODULE_LINK ${MODULE_LINK}
+        gcrashpad
+    )
+
+endif(BUILD_CRASHPAD_CLIENT)
+# ----------------
 
 include(${PROJECT_SOURCE_DIR}/build/module.cmake)
 
+# === Install crashpad handler ===
+set(CPAD_ROOT_PATH ${CMAKE_CURRENT_LIST_DIR}/crashpad_handler)
+if (OS_IS_LIN)
+    install(PROGRAMS ${CPAD_ROOT_PATH}/linux/x86-64/crashpad_handler DESTINATION bin)
+elseif(OS_IS_WIN)
+    install(PROGRAMS ${CPAD_ROOT_PATH}/windows/x86-64/crashpad_handler.exe DESTINATION bin)
+elseif(OS_IS_MAC)
+    install(PROGRAMS ${CPAD_ROOT_PATH}/macos/x86-64/crashpad_handler DESTINATION ${Mscore_SHARE_NAME}Contents/MacOS)
+endif()
