--- src/framework/telemetry/qml/MuseScore/Telemetry/qmldir
@@ -1,1 +1,0 @@
-module MuseScore.Telemetry
