--- src/framework/telemetry/telemetry.qrc
@@ -1,3 +1,0 @@
-<RCC>
-    <qresource prefix="/"/>
-</RCC>
